Refactored default:scc_inventory_enabled to scc:compliance_checks_enabled
No backwards compatibility was added here since the module has not been
"released".

With this change you no longer have to explicitly define a class for compliance
checks to be active. This change also provides a new class that can disable
compliance checks.

Notably, this change moved the class to the scc namespace.